Post code review amendments and refactor
